requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile '.\requirements.in'
#
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.56.0
    # via matplotlib
joblib==1.4.2
    # via
    #   -r .\requirements.in
    #   scikit-learn
kiwisolver==1.4.8
    # via matplotlib
matplotlib==3.10.1
    # via -r .\requirements.in
numpy==2.2.3
    # via
    #   -r .\requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   xgboost
packaging==24.2
    # via matplotlib
pandas==2.2.3
    # via -r .\requirements.in
pillow==11.1.0
    # via matplotlib
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r .\requirements.in
pytz==2025.1
    # via pandas
scikit-learn==1.6.1
    # via -r .\requirements.in
scipy==1.15.2
    # via
    #   -r .\requirements.in
    #   scikit-learn
    #   xgboost
six==1.17.0
    # via python-dateutil
threadpoolctl==3.6.0
    # via scikit-learn
tzdata==2025.1
    # via pandas
xgboost==3.0.0
    # via -r .\requirements.in
